Third question a answered

+++ src/main/sql/schema.sql
@@ -1,7 +1,7 @@
 CREATE table Product (
     id int primary key,
     name varchar(100),
-    price int,
+    price float,
     creation_datetime timestamp default current_date
 );
